// PickEmLeagueFrontEnd/ClientApp/src/Apis/models/SeasonSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PickEmLeague
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import { User, UserFromJSON, UserFromJSONTyped, UserToJSON } from "./";

/**
 *
 * @export
 * @interface SeasonSummary
 */
export interface SeasonSummary {
  /**
   *
   * @type {User}
   * @memberof SeasonSummary
   */
  user?: User;
  /**
   *
   * @type {number}
   * @memberof SeasonSummary
   */
  score?: number;
  /**
   *
   * @type {number}
   * @memberof SeasonSummary
   */
  correctPicks?: number;
  /**
   *
   * @type {number}
   * @memberof SeasonSummary
   */
  place?: number;
}

export function SeasonSummaryFromJSON(json: any): SeasonSummary {
  return SeasonSummaryFromJSONTyped(json, false);
}

export function SeasonSummaryFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): SeasonSummary {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    user: !exists(json, "user") ? undefined : UserFromJSON(json["user"]),
    score: !exists(json, "score") ? undefined : json["score"],
    correctPicks: !exists(json, "correctPicks")
      ? undefined
      : json["correctPicks"],
    place: !exists(json, "place") ? undefined : json["place"],
  };
}

export function SeasonSummaryToJSON(value?: SeasonSummary | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    user: UserToJSON(value.user),
    score: value.score,
    correctPicks: value.correctPicks,
    place: value.place,
  };
}